Implementation of a 2 Lock concurrent queue -- NOT TESTED
The queue is based on Michael & Scott's alogrithm in their paper and
provides basic support for Queue<E> interface.  Collection<E> is NOT
supported but can be added later (if it makes sense -- doesn't to me).

Obviously the Lea non-blocking renditions of the algos in the same
paper are superior.  Main interest here was (a) to build it, and (b)
to use it in situations where a fan-in : Q : fan-out strucutre is
inherent.  In those cases, the contention is 1:1 and we can do away
with the locks for Head and Tail so it is more efficient.